Remove unreachable legacy features
