+++ Foundation Lab/Models/DataModels.swift
@@ -50,20 +50,3 @@ struct ChatMessage: Identifiable, Equatable {
         self.isContextSummary = isContextSummary
     }
 }
-
-// MARK: - Request/Response Models
-
-struct RequestResponsePair: Identifiable {
-    let id = UUID()
-    let request: String
-    let response: String
-    let isError: Bool
-    let timestamp: Date
-
-    init(request: String, response: String, isError: Bool = false) {
-        self.request = request
-        self.response = response
-        self.isError = isError
-        self.timestamp = Date()
-    }
-}
