--- Foundation Lab/Views/Examples/BasicChatView.swift
@@ -1,119 +1,0 @@
-//
-//  BasicChatView.swift
-//  FoundationLab
-//
-//  Created by Rudrank Riyam on 6/29/25.
-//
-
-import FoundationLabCore
-import SwiftUI
-
-struct BasicChatView: View {
-    @State private var currentPrompt = FoundationLabExampleDemo.basicChat.defaultPrompt
-    @State private var instructions = FoundationLabExampleDemo.basicChat.defaultSystemPrompt ?? ""
-    @State private var executor = ExampleExecutor()
-    @State private var showInstructions = false
-    @State private var usePermissiveGuardrails = false
-
-    var body: some View {
-        ExampleViewBase(
-            title: "One-shot",
-            description: "Single prompt-response interaction with the AI assistant",
-            defaultPrompt: FoundationLabExampleDemo.basicChat.defaultPrompt,
-            currentPrompt: $currentPrompt,
-            isRunning: executor.isRunning,
-            errorMessage: executor.errorMessage,
-            codeExample: DefaultPrompts.basicChatCode(
-                prompt: currentPrompt,
-                instructions: showInstructions && !instructions.isEmpty ? instructions : nil
-            ),
-            onRun: executeChat,
-            onReset: resetToDefaults
-        ) {
-            VStack(spacing: Spacing.large) {
-                // Instructions Section
-                VStack(alignment: .leading, spacing: 0) {
-                    Button(action: { showInstructions.toggle() }, label: {
-                        HStack(spacing: Spacing.small) {
-                            Image(systemName: showInstructions ? "chevron.down" : "chevron.right")
-                                .font(.caption2)
-                                .foregroundColor(.secondary)
-
-                            Text("Instructions")
-                                .font(.callout)
-                                .foregroundColor(.primary)
-
-                            Spacer()
-                        }
-                    })
-                    .buttonStyle(.plain)
-                    .padding(.horizontal, Spacing.medium)
-                    .padding(.vertical, Spacing.small)
-
-                    if showInstructions {
-                        VStack(alignment: .leading, spacing: Spacing.small) {
-                            TextEditor(text: $instructions)
-                                .font(.body)
-                                .scrollContentBackground(.hidden)
-                                .padding(Spacing.medium)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                                .background(Color.gray.opacity(0.05))
-                                .cornerRadius(8)
-                        }
-                        .padding(.horizontal, Spacing.medium)
-                        .padding(.bottom, Spacing.small)
-                        .transition(.opacity.combined(with: .move(edge: .top)))
-                    }
-                }
-                .background(Color.gray.opacity(0.1))
-                .cornerRadius(12)
-
-                // Guardrails Toggle
-                PermissiveGuardrailsToggle(isEnabled: $usePermissiveGuardrails)
-
-                // Prompt Suggestions
-                PromptSuggestions(
-                    suggestions: FoundationLabExampleDemo.basicChat.suggestions,
-                    onSelect: { currentPrompt = $0 }
-                )
-
-                // Prompt History
-                if !executor.promptHistory.isEmpty {
-                    PromptHistory(
-                        history: executor.promptHistory,
-                        onSelect: { currentPrompt = $0 }
-                    )
-                }
-
-                // Result Display
-                if !executor.result.isEmpty {
-                    ResultDisplay(
-                        result: executor.result,
-                        isSuccess: executor.errorMessage == nil,
-                        tokenCount: executor.lastTokenCount
-                    )
-                }
-            }
-        }
-    }
-
-    private func executeChat() {
-        Task {
-            let guardrails: FoundationLabGuardrails = usePermissiveGuardrails ?
-                .permissiveContentTransformations : .default
-
-            await executor.executeBasic(
-                prompt: currentPrompt,
-                instructions: instructions.isEmpty ? nil : instructions,
-                guardrails: guardrails
-            )
-        }
-    }
-
-    private func resetToDefaults() {
-        currentPrompt = "" // Clear the prompt completely
-        instructions = FoundationLabExampleDemo.basicChat.defaultSystemPrompt ?? ""
-        usePermissiveGuardrails = false
-        executor.clearAll() // Clear all results, errors, and history
-    }
-}
