--- Foundation Lab/Views/Examples/CreativeWritingView.swift
@@ -1,99 +1,0 @@
-//
-//  CreativeWritingView.swift
-//  FoundationLab
-//
-//  Created by Rudrank Riyam on 6/29/25.
-//
-
-import FoundationLabCore
-import SwiftUI
-
-struct CreativeWritingView: View {
-  @State private var currentPrompt = FoundationLabExampleDemo.creativeWriting.defaultPrompt
-  @State private var instructions = FoundationLabExampleDemo.creativeWriting.defaultSystemPrompt ?? ""
-  @State private var executor = ExampleExecutor()
-  @State private var showInstructions = false
-
-  var body: some View {
-    ExampleViewBase(
-      title: "Creative Writing",
-      description: "Generate stories, poems, and creative content",
-      defaultPrompt: FoundationLabExampleDemo.creativeWriting.defaultPrompt,
-      currentPrompt: $currentPrompt,
-      isRunning: executor.isRunning,
-      errorMessage: executor.errorMessage,
-      codeExample: DefaultPrompts.creativeWritingCode(prompt: currentPrompt,
-                                                        instructions: showInstructions && !instructions.isEmpty ?
-                                                                     instructions : nil),
-      onRun: executeCreativeWriting,
-      onReset: resetToDefaults
-    ) {
-      VStack(spacing: 16) {
-        // Info Banner
-        HStack {
-          Image(systemName: "info.circle")
-            .foregroundColor(.blue)
-          Text("Creates structured story outlines with plot, characters, and themes")
-            .font(.caption)
-            .foregroundColor(.secondary)
-          Spacer()
-        }
-        .padding()
-        .background(Color.blue.opacity(0.1))
-        .cornerRadius(8)
-
-        // Prompt Suggestions
-        PromptSuggestions(
-          suggestions: FoundationLabExampleDemo.creativeWriting.suggestions,
-          onSelect: { currentPrompt = $0 }
-        )
-
-        // Prompt History
-        if !executor.promptHistory.isEmpty {
-          PromptHistory(
-            history: executor.promptHistory,
-            onSelect: { currentPrompt = $0 }
-          )
-        }
-
-        // Result Display
-        if !executor.result.isEmpty {
-          VStack(alignment: .leading, spacing: 12) {
-            Label("Story Outline", systemImage: "book.closed")
-              .font(.headline)
-
-            ResultDisplay(
-              result: executor.result,
-              isSuccess: executor.errorMessage == nil,
-              tokenCount: executor.lastTokenCount
-            )
-          }
-        }
-      }
-    }
-  }
-
-  private func executeCreativeWriting() {
-    Task {
-      await executor.executeStructured(
-        prompt: currentPrompt,
-        type: StoryOutline.self,
-        instructions: instructions.isEmpty ? nil : instructions
-      ) { story in
-        story.plainTextSummary
-      }
-    }
-  }
-
-  private func resetToDefaults() {
-    currentPrompt = "" // Clear the prompt completely
-    instructions = FoundationLabExampleDemo.creativeWriting.defaultSystemPrompt ?? ""
-    executor.clearAll() // Clear all results, errors, and history
-  }
-}
-
-#Preview {
-  NavigationStack {
-    CreativeWritingView()
-  }
-}
