--- Foundation Lab/Views/Examples/ExamplesView.swift
@@ -1,20 +1,0 @@
-//
-//  ExamplesView.swift
-//  FoundationLab
-//
-//  Created by Rudrank Riyam on 6/15/25.
-//
-
-import SwiftUI
-import FoundationModels
-
-struct ExamplesView: View {
-    var body: some View {
-        HomeView()
-    }
-}
-
-#Preview {
-    ExamplesView()
-        .environment(NavigationCoordinator.shared)
-}
